% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{temp_se_slope}
\alias{temp_se_slope}
\title{Standard error of the slope of an OLS fit of \code{y} on a single predictor}
\usage{
temp_se_slope(y, x)
}
\arguments{
\item{y}{Response vector.}

\item{x}{Predictor vector.}
}
\value{
The slope standard error.
}
\description{
Standard error of the slope of an OLS fit of \code{y} on a single predictor
}
