% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{temp_slope}
\alias{temp_slope}
\title{Slope of an OLS fit of \code{y} on a single predictor}
\usage{
temp_slope(y, x)
}
\arguments{
\item{y}{Response vector.}

\item{x}{Predictor vector.}
}
\value{
The slope coefficient.
}
\description{
Slope of an OLS fit of \code{y} on a single predictor
}
