% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset.R
\name{weibull_offset}
\alias{weibull_offset}
\title{Weibull offset (upper limit) for a vector of day-of-season values}
\usage{
weibull_offset(days)
}
\arguments{
\item{days}{Numeric vector of days (e.g. days after 1 July).}
}
\value{
Estimated end (offset) day.
}
\description{
Weibull offset (upper limit) for a vector of day-of-season values
}
