% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset.R
\name{weibull_onset}
\alias{weibull_onset}
\title{Weibull onset (lower limit) for a vector of day-of-season values}
\usage{
weibull_onset(days)
}
\arguments{
\item{days}{Numeric vector of days (e.g. days after 1 July).}
}
\value{
Estimated onset day.
}
\description{
Weibull onset (lower limit) for a vector of day-of-season values
}
